Move throwIf and throwUnless from Obj to Ex
