+++ README.md
@@ -82,18 +82,6 @@ private InputStream openFileOrResource(String name) {
 }
 ```
 
-#### `throwIf` and `throwUnless`
-
-Throws a provided exception if the given `value` satisfies (or doesn't satisfy) the provided `predicate`.
-e.g.,
-
-```java
-public Map<String, String> loadProperties() {
-    return Obj.throwIf(Properties.loadFromJson(FILENAME), Map::isEmpty,
-            () -> new IllegalStateException("Properties must not be empty"));
-}
-```
-
 #### `newInstanceOf`
 
 Creates a new instance of the same type as the input object if possible, otherwise, returns empty. e.g.,
@@ -159,6 +147,18 @@ public Document parseXml(String pathname) throws XmlProcessingException {
                     .newDocumentBuilder()       // throws ParserConfigurationException
                     .parse(new File(pathname)), // throws SAXException, IOException
             XmlProcessingException::new);
+}
+```
+
+#### `throwIf` and `throwUnless`
+
+Throws a provided exception if the given `value` satisfies (or doesn't satisfy) the provided `predicate`.
+e.g.,
+
+```java
+public Map<String, String> loadProperties() {
+    return Ex.throwIf(Properties.loadFromJson(FILENAME), Map::isEmpty,
+            () -> new IllegalStateException("Properties must not be empty"));
 }
 ```
