+++ src/main/java/io/blt/util/Ex.java
@@ -27,6 +27,8 @@ package io.blt.util;
 import io.blt.util.functional.ThrowingRunnable;
 import io.blt.util.functional.ThrowingSupplier;
 import java.util.function.Function;
+import java.util.function.Predicate;
+import java.util.function.Supplier;
 
 /**
  * Static utility methods centred around {@code Exception} and {@code Throwable}.
@@ -105,6 +107,57 @@ public final class Ex {
             runnable.run();
             return null;
         }, transformer);
+    }
+
+    /**
+     * Throws the specified {@code throwable} if the given {@code value} satisfies the provided {@code predicate}.
+     * For convenience, {@code value} is returned.
+     * e.g.,
+     * <pre>{@code
+     * public Map<String, String> loadProperties() {
+     *     return throwIf(Properties.loadFromJson(FILENAME), Map::isEmpty,
+     *             () -> new IllegalStateException("Properties must not be empty"));
+     * }
+     * }</pre>
+     *
+     * @param value     the value to be checked
+     * @param predicate the predicate to be evaluated
+     * @param throwable the supplier for the throwable to be thrown
+     * @param <T>       the type of the value
+     * @param <E>       the type of the throwable
+     * @return {@code value}
+     * @throws E if the given {@code value} satisfies the provided {@code predicate}
+     * @see Ex#throwUnless(Object, Predicate, Supplier)
+     */
+    public static <T, E extends Throwable> T throwIf(
+            T value, Predicate<? super T> predicate, Supplier<? extends E> throwable) throws E {
+        if (predicate.test(value)) {
+            throw throwable.get();
+        }
+        return value;
+    }
+
+    /**
+     * Throws the specified {@code throwable} if the given {@code value} does not satisfy the provided {@code predicate}.
+     * For convenience, {@code value} is returned.
+     * e.g.,
+     * <pre>{@code
+     * throwUnless(properties, p -> p.containsKey("host"),
+     *         () -> new IllegalStateException("Properties must contain a host"));
+     * }</pre>
+     *
+     * @param value     the value to be checked
+     * @param predicate the predicate to be evaluated
+     * @param throwable the supplier for the throwable to be thrown
+     * @param <T>       the type of the value
+     * @param <E>       the type of the throwable
+     * @return {@code value}
+     * @throws E if the given {@code value} does not satisfy the provided {@code predicate}
+     * @see Ex#throwIf(Object, Predicate, Supplier)
+     */
+    public static <T, E extends Throwable> T throwUnless(
+            T value, Predicate<? super T> predicate, Supplier<? extends E> throwable) throws E {
+        return throwIf(value, predicate.negate(), throwable);
     }
 
 }
