+++ src/main/java/io/blt/util/Obj.java
@@ -140,54 +140,25 @@ public final class Obj {
     }
 
     /**
-     * Throws the specified {@code throwable} if the given {@code value} satisfies the provided {@code predicate}.
-     * For convenience, {@code value} is returned.
-     * e.g.,
-     * <pre>{@code
-     * public Map<String, String> loadProperties() {
-     *     return throwIf(Properties.loadFromJson(FILENAME), Map::isEmpty,
-     *             () -> new IllegalStateException("Properties must not be empty"));
-     * }
-     * }</pre>
-     *
-     * @param value     the value to be checked
-     * @param predicate the predicate to be evaluated
-     * @param throwable the supplier for the throwable to be thrown
-     * @param <T>       the type of the value
-     * @param <E>       the type of the throwable
-     * @return {@code value}
-     * @throws E if the given {@code value} satisfies the provided {@code predicate}
-     * @see Obj#throwUnless(Object, Predicate, Supplier)
+     * @deprecated
+     * This has been moved to {@link Ex}.
+     * <p>Use {@link Ex#throwIf(Object, Predicate, Supplier)} instead.</p>
      */
+    @Deprecated(since = "1.0.7", forRemoval = true)
     public static <T, E extends Throwable> T throwIf(
             T value, Predicate<? super T> predicate, Supplier<? extends E> throwable) throws E {
-        if (predicate.test(value)) {
-            throw throwable.get();
-        }
-        return value;
+        return Ex.throwIf(value, predicate, throwable);
     }
 
     /**
-     * Throws the specified {@code throwable} if the given {@code value} does not satisfy the provided {@code predicate}.
-     * For convenience, {@code value} is returned.
-     * e.g.,
-     * <pre>{@code
-     * throwUnless(properties, p -> p.containsKey("host"),
-     *         () -> new IllegalStateException("Properties must contain a host"));
-     * }</pre>
-     *
-     * @param value     the value to be checked
-     * @param predicate the predicate to be evaluated
-     * @param throwable the supplier for the throwable to be thrown
-     * @param <T>       the type of the value
-     * @param <E>       the type of the throwable
-     * @return {@code value}
-     * @throws E if the given {@code value} does not satisfy the provided {@code predicate}
-     * @see Obj#throwIf(Object, Predicate, Supplier)
+     * @deprecated
+     * This has been moved to {@link Ex}.
+     * <p>Use {@link Ex#throwUnless(Object, Predicate, Supplier)} instead.</p>
      */
+    @Deprecated(since = "1.0.7", forRemoval = true)
     public static <T, E extends Throwable> T throwUnless(
             T value, Predicate<? super T> predicate, Supplier<? extends E> throwable) throws E {
-        return throwIf(value, predicate.negate(), throwable);
+        return Ex.throwUnless(value, predicate, throwable);
     }
 
     /**
